Refactoring changes to keep things clean.
