+++ docker-compose.yml
@@ -1,7 +1,7 @@
 version: "3.8"
 
 networks:
-  monitoring:
+  periodum:
     driver: bridge
 
 volumes:
@@ -25,15 +25,15 @@ services:
       - db
     restart: unless-stopped
     networks:
-      - monitoring
+      - periodum
   db:
     image: "mongo:${MONGO_DB_VERSION}"
     container_name: periodum_db
     volumes:
-        - ./db/mongodb:/data/db
+      - ./db/mongodb:/data/db
     restart: unless-stopped
     networks:
-      - monitoring
+      - periodum
   grafana:
     image: grafana/grafana:8.2.0
     container_name: grafana
@@ -46,11 +46,9 @@ services:
       - GF_AUTH_ANONYMOUS_ORG_ROLE=Admin
       - GF_SERVER_HTTP_PORT=8080
     ports:
-      - "8080:8080"
-    expose:
-      - 8080
+      - "127.0.0.1:8080:8080"
     networks:
-      - monitoring
+      - periodum
   node-exporter:
     image: prom/node-exporter:latest
     container_name: node-exporter
@@ -65,11 +63,9 @@ services:
       - '--path.sysfs=/host/sys'
       - '--collector.filesystem.mount-points-exclude=^/(sys|proc|dev|host|etc)($$|/)'
     ports:
-      - "9100:9100"
-    expose:
-      - 9100
+      - "127.0.0.1:9100:9100"
     networks:
-      - monitoring
+      - periodum
   prometheus:
     image: prom/prometheus:latest
     container_name: prometheus
@@ -84,8 +80,6 @@ services:
       - '--web.console.templates=/etc/prometheus/consoles'
       - '--web.enable-lifecycle'
     ports:
-      - "9090:9090"
-    expose:
-      - 9090
+      - "127.0.0.1:9090:9090"
     networks:
-      - monitoring
+      - periodum
