+++ package.json
@@ -6,7 +6,7 @@
     "lint": "eslint . --ext .ts",
     "lint-fix": "eslint . --ext .ts --fix",
     "build": "swc src -d dist",
-    "build-dev": "npm run build && NODE_ENV=development node --inspect dist/app.js",
+    "build-dev": "npm run build && NODE_ENV=development node --inspect --experimental-specifier-resolution=node dist/app.js",
     "dev": "nodemon --ext ts --exec \"npm run build-dev\"",
     "start": "npm run build && NODE_ENV=production node dist/app.js",
     "test": "NODE_ENV=test mocha --check-leaks -r ts-node/register -r tsconfig-paths/register \"test/**/*.spec.ts\"",
